+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/OrchestrationClient.java
@@ -344,6 +344,24 @@ public class OrchestrationClient {
   }
 
   /**
+   * Create a new orchestration client with multiple custom headers added to every call made with
+   * this client
+   *
+   * @param headers a map of key value pairs for the custom headers to add
+   * @return a new client.
+   * @since 1.22.0
+   */
+  @Nonnull
+  public OrchestrationClient withHeaders(@Nonnull final Map<String, String> headers) {
+    final var newClient = new OrchestrationClient(this.executor);
+    newClient.customHeaders.addAll(this.customHeaders);
+    headers.entrySet().stream()
+        .map(e -> new Header(e.getKey(), e.getValue()))
+        .forEach(newClient.customHeaders::add);
+    return newClient;
+  }
+
+  /**
    * Create a new orchestration client for the given resource group and scenario.
    *
    * @param resourceGroup the resource group